character stat 수정
파일에서 읽은 데이터로 사용

+++ Server/Server/Source/Room/Room.cpp
@@ -41,8 +41,8 @@ void Room::Init(int id, int teamLifeCount, int playerLimit)
 	mTeams.clear();
 
 	// team game
-	mTeams[TeamCode::RED].Init(teamLifeCount);
-	mTeams[TeamCode::BLUE].Init(teamLifeCount);
+	mTeams[(int)TeamCode::RED].Init(teamLifeCount);
+	mTeams[(int)TeamCode::BLUE].Init(teamLifeCount);
 
 	InitRoomCode();
 }
